Rename component sage-avatar
