+++ src/app/app.module.ts
@@ -18,7 +18,7 @@ import { FiltersModule } from './components/filters/filters.module';
 import { AppConfigService } from './app-config.service';
 import { AppConfig, APP_CONFIG } from './app.config';
 import { PageTitleModule } from './components/page-title/page-title.module';
-import { UserAvatarModule } from './components/user-avatar/user-avatar.module';
+import { AvatarModule } from './components/avatar/avatar.module';
 
 export function apiConfigFactory(): Configuration {
   const params: ConfigurationParameters = {
@@ -39,7 +39,7 @@ export function apiConfigFactory(): Configuration {
     DatabaseSeedModule,
     FiltersModule,
     PageTitleModule,
-    UserAvatarModule
+    AvatarModule
   ],
   declarations: [AppComponent],
   providers: [
