+++ src/app/components/avatar/avatar.component.html
