+++ src/app/components/avatar/avatar.component.ts
@@ -1,0 +1,14 @@
+import { Component, Input, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'sage-avatar',
+  templateUrl: './avatar.component.html',
+  styleUrls: ['./avatar.component.scss'],
+})
+export class AvatarComponent {
+  @Input() name = '';
+  @Input() avatarUrl = '';
+  @Input() size = 40;
+
+  constructor() {}
+}
