+++ src/app/components/avatar/avatar.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { AvatarModule as NgxAvatarModule } from 'ngx-avatar';
+import { AvatarComponent } from './avatar.component';
+
+@NgModule({
+  imports: [CommonModule, NgxAvatarModule],
+  declarations: [AvatarComponent],
+  exports: [AvatarComponent],
+})
+export class AvatarModule {}
