+++ src/app/pages/homepage/homepage.component.html
@@ -7,8 +7,8 @@
       [routerLink]="[username]">Update Profile</button>
     Configuration: {{appConfig | json}}
 
-    <sage-user-avatar [name]="user.name" [avatarUrl]="user.avatarUrl" [size]=100></sage-user-avatar>
-    <sage-user-avatar [name]="user.name" [size]=60></sage-user-avatar>
+    <sage-avatar [name]="user.name" [avatarUrl]="user.avatarUrl" [size]=100></sage-avatar>
+    <sage-avatar [name]="user.name" [size]=60></sage-avatar>
   </div>
 </main>
